InterfaceVisualisation: renamed to InterfaceController and moved out of UI namespace
