--- Assets/Scripts/UI.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: d75b9cad24b447868c52f5c7d1ab0303
-timeCreated: 1695107991
